move context to a folder
